原生支付（扫码支付）V3版
修复 元 转 分的精度问题

+++ native_v3.php
@@ -82,8 +82,8 @@ class WxpayService
             'notify_url' => $this->notifyUrl,       //通知URL必须为直接可访问的URL，不允许携带查询串。
             'out_trade_no' => $this->outTradeNo,      //商户系统内部订单号，只能是数字、大小写字母_-*且在同一个商户号下唯一，详见【商户订单号】。特殊规则：最小字符长度为6
             'amount'=>array(
-                'total'=>intval($this->totalFee * 100), //订单总金额，单位为分
-                'currency'=>'CNY', //CNY：人民币，境内商户号仅支持人民币
+                'total'=> floatval($this->totalFee) * 100, //订单总金额，单位为分
+                'currency'=> 'CNY', //CNY：人民币，境内商户号仅支持人民币
             ),
             'scene_info'=>array(        //支付场景描述
                 'payer_client_ip'=>'127.0.0.1'   //调用微信支付API的机器IP
